+++ suraft/src/type_config/async_runtime/tokio_impls/tokio_runtime.rs
@@ -77,7 +77,7 @@ impl AsyncRuntime for TokioRuntime {
 
     #[inline]
     fn thread_rng() -> Self::ThreadLocalRng {
-        rand::thread_rng()
+        rand::rng()
     }
 
     type Mpsc = mpsc_impl::TokioMpsc;
